feat(Signup): Implement new Signup component with Firebase authentication and user data storage
